+++ forum/templates/forum/email/new_follower.html
@@ -1,0 +1,26 @@
+{% load i18n cm_tags %}
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<title>{% translate 'Verify Email address' %}</title>
+		{% inline_css "cm_main/css/email_msg_css.css" %}
+	</head>
+	<body>
+		<div class="container">
+			<div class="green">
+				<h1 class="center" style="padding:2px">
+					{%blocktranslate asvar trans_title with post_title=post.title %}Your post '{{post_title}}' has a new follower!{%endblocktranslate%}
+					{% title trans_title%}
+				</h1>
+			</div>
+			<div class="center">
+				<p class="mt-2">
+					{%blocktranslate with follower_name=follower.get_full_name post_title=post.title site_name=settings.SITE_NAME%}{{follower_name}} is now following your post <a href="{{post_url}}">"{{post_title}}"</a> on {{site_name}}{%endblocktranslate%}
+				</p>
+			</div>
+			<p class="small mt-3 center">{% trans "Do not reply to this mail, it is machine generated." %}</p>
+		</div>
+	</body>
+</html>
